+++ apps/avifgainmaputil/swapbase_command.cc
@@ -166,6 +166,8 @@ avifResult SwapBaseCommand::Run() {
   decoder->maxThreads = arg_jobs_.jobs.value();
   decoder->imageContentToDecode |= AVIF_IMAGE_CONTENT_GAIN_MAP;
   decoder->ignoreICC = arg_image_read_.ignore_profile;
+  decoder->ignoreExif = arg_image_read_.ignore_exif;
+  decoder->ignoreXMP = arg_image_read_.ignore_xmp;
   avifResult result = ReadAvif(decoder.get(), arg_input_filename_);
   if (result != AVIF_RESULT_OK) {
     return result;